+++ src/main/java/ca/appsimulations/jlqninterface/lqn/entities/EntryMakingCallType.java
@@ -1,32 +1,33 @@
 package ca.appsimulations.jlqninterface.lqn.entities;
+
 import ca.appsimulations.jlqninterface.lqn.model.LqnModel;
 
 /**
  * @author Yasir Shoaib (2011,2012) Contributors: Yasir Shoaib - Implementation
- * 
- *         Some LQN classes and their members are outlined as UML class diagrams
- *         in LQNS User Manual. For details regarding these LQN classes and
- *         members refer to LQNS User Manual.
+ * <p>
+ * Some LQN classes and their members are outlined as UML class diagrams
+ * in LQNS User Manual. For details regarding these LQN classes and
+ * members refer to LQNS User Manual.
  */
 
 public abstract class EntryMakingCallType extends MakingCallType {
-	protected float prob;
+    protected double prob;
 
-	public EntryMakingCallType(LqnModel lqnModel, String strDestEntry, float prob) {
-		super(lqnModel, strDestEntry);
-		this.prob = prob;
-	}
+    public EntryMakingCallType(LqnModel lqnModel, String strDestEntry, double prob) {
+        super(lqnModel, strDestEntry);
+        this.prob = prob;
+    }
 
-	public EntryMakingCallType(LqnModel lqnModel, Entry destEntry, float prob) {
-		super(lqnModel, destEntry);
-		this.prob = prob;
-	}
+    public EntryMakingCallType(LqnModel lqnModel, Entry destEntry, double prob) {
+        super(lqnModel, destEntry);
+        this.prob = prob;
+    }
 
-	public float getProb() {
-		return prob;
-	}
+    public double getProb() {
+        return prob;
+    }
 
-	public void setProb(float prob) {
-		this.prob = prob;
-	}
+    public void setProb(double prob) {
+        this.prob = prob;
+    }
 }
